Finishing up drive train
The drive train that was previously worked on is now in its completed state and ready to be looked over because it is no longer incomplete because I Cor y has finished it and will be yelled at when I do something wrong because my completed state is not up to the standards of Clinton even though this is finished because it is a completed state and is completed. In short it is very much completed.